Make sure network does come up correctly during reboot
OpenVSwitch when started during boot will automatically recreate bridges and
add the physical interfaces according to its configuration database. It will
however not automatically reassign the IP addresses from the physical interface
to the internal bridge port. As a result of this network connectivity on that
interface is lost.  This is especially problematic when the physical interface
that hosts the "admin" network (ususally eth0) is enslaved into one of the ovs
bridges (e.g. 'br-fixed' when using an openvswitch with the vlan driver). In
that case the node will loose connectivity to the admin node and the initial
chef-client, which would usually fix the IP assignments will fail.

By creating and enabling init scripts that transfer the IP addresses from the
physical (or bond) interfaces to the correct ovs bridge we make sure that the
node's network is working even before chef-client ran for the first time.

This should also help us to remove the requirement for a separate physical
interface for the "nova_fixed" network when using openvswitch with VLANs.
(See http://github.com/crowbar/barclamp-quantum/pull/113)

Addresses: https://bugzilla.novell.com/show_bug.cgi?id=825644