Fix check of ha enablement of l3 agent
When neutron-server and nova-compute are deployed on the same host (non-HA),
while neutron-l3 is deployed on separate machines in a HA cluster the check as
it was before the fix would falsely assume HA mode for the openvswitch-agent on
the compute nodes. As a result the agent is not started there.

Fixes: https://bugzilla.novell.com/show_bug.cgi?id=885832
(cherry picked from commit c87ad4228765222bf0c2bf4b12b661c8101670b1)