Allow to set use_external_dhcp
This patch makes possible to set the flag use_external_dhcp.
If use_external_dhcp is true a metadata proxy will be spawned by
the DHCP agent even if the subnet has dhcp_enable set to False.
This is useful for cloud where an external appliance is used to
provide DHCP services.