Create "floating" network as "flat" provider network (bsc#946882)
This should work regardless of the mechanism driver that is used  (openvswitch
or linuxbridge) as we have the ml2 type driver "flat" enabled unconditionally.
This is also in line with the behaviour of devstack.

For linuxbridge setups this commit also adjust the interface mappings for
external networks to use the VLAN interface instead for the underlying physical
interface. That's required for "flat" networks.

We also make sure to specify the right "--provider:physical_network" parameter
for the floating network. Otherwise the floating network might endup on the
wrong phyiscal interfaces in deployments that use different network conduits
for "nova_floating" and "nova_fixed".