Fix database attributes of the floating network and its router port (bsc#967009)
Because of several bugs (at least bsc#946882 and bsc#955811) we have setups
where the floating network was created with the wrong provider network
settings. While this somehow continued to work and got unnoticed it will
create problems on systems upgrading from juno to liberty. This fix tries to
to re-adjust the attributes directly neutron database to their correct values
to avoid problems during the upgrade.

https://bugzilla.suse.com/show_bug.cgi?id=967009